Library

Library

Public Documentation

Internal Documentation

Index

Public Documentation

MinFEMFlow.MinFEMFlow — Module

MinFEMFlow

A minimalistic Stokes flow solver in julia based on MinFEM.

MinFEMFlow.FlowSolver — Type

mutable struct FlowSolver

Stores all data that is static throughout iteration of the interior-point method. Is supposed to be generated by corresponding constructor.

Fields

mesh::MinFEM.Mesh: Meshed computational domain

inflow_boundary::Set{Int64}: Inflow boundary with prescribed Dirichlet velocity

noslip_boundary::Set{Int64}: No-slip boundary with prescribed Dirichlet values

inflow::Vector{Float64}: Inflow velocity function

verbose::Bool: Switch for writing output during solve

solve::Function: Flow solution routine inherently determining the flow type

v::Union{Missing, Vector{Float64}}: Resulting velocity field

p::Union{Missing, Vector{Float64}}: Resulting pressure

time::Union{Missing, Float64}: Required solution time also validating if a result is present

MinFEM.solve! — Method

solve!(fs::FlowSolver) -> Any

Solves flow by calling the internal routine.

MinFEM.write_to_vtk — Method

write_to_vtk(
    fs::FlowSolver,
    file_name::String
) -> Vector{String}

Writes velocity and pressure of the solved flow to a .vtu-file under the given file_name.

MinFEMFlow.StokesFlowSolver — Method

StokesFlowSolver(
    mesh::Mesh,
    inflow_boundary::Union{Set{Boundary}, Set{Int64}},
    noslip_boundary::Union{Set{Boundary}, Set{Int64}},
    inflow::Union{Vector{Float64}, Function};
    verbose::Bool = true
) -> FlowSolver

Returns FlowSolver object representing the solution routine for a Stokes flow.

MinFEMFlow.assemble_saddlepointmatrix_stokes — Method

assemble_saddlepointmatrix_stokes(
    mesh::MinFEM.Mesh;
    alpha
) -> SparseArrays.SparseMatrixCSC{Float64, Int64}

Returns saddle point system matrix including inf-sup stabilization for the solution of the discretized Stokes equation.

MinFEMFlow.copy_result! — Method

copy_result!(
    target::FlowSolver,
    origin::FlowSolver
) -> Float64

Copies result from one flow solver to another. Does not check if problems in the solvers match.

MinFEMFlow.invalidate! — Method

invalidate!(fs::FlowSolver) -> Missing

Removes result, i.e. velocity, pressure and time, from given FlowSolver.

MinFEMFlow.set_boundary_inflow! — Method

set_boundary_inflow!(
    fs::FlowSolver,
    boundary::Union{Set{Int64}, Set{MinFEM.Boundary}}
) -> Missing

Updates inflow boundary and then invalidates the solution.

MinFEMFlow.set_boundary_noslip! — Method

set_boundary_noslip!(
    fs::FlowSolver,
    boundary::Union{Set{Int64}, Set{MinFEM.Boundary}}
) -> Missing

Updates no slip boundary and then invalidates the solution.

MinFEMFlow.set_inflow! — Method

set_inflow!(
    fs::FlowSolver,
    inflow::Union{Function, Vector{Float64}}
) -> Missing

Updates inflow function and then invalidates the solution.

MinFEMFlow.set_mesh! — Method

set_mesh!(fs::FlowSolver, mesh::MinFEM.Mesh) -> Missing

Updates mesh and then invalidates the solution.

MinFEMFlow.set_result! — Method

set_result!(
    fs::FlowSolver,
    v::Vector{Float64},
    p::Vector{Float64},
    t::Float64
) -> Float64

Set results for a FlowSolver. Does not check if solution matches the problem.

MinFEMFlow.set_verbose! — Method

set_verbose!(fs::FlowSolver, verbose::Bool) -> Bool

Updates wether output shall be written to console.

Internal Documentation

MinFEMFlow.assemble_incompressibility_stokes — Method

assemble_incompressibility_stokes(
    mesh::MinFEM.Mesh
) -> SparseArrays.SparseMatrixCSC{Float64, Int64}

Returns the matrix for a linear incompressibility block in the saddle point system for the discretized Stokes equation.

MinFEMFlow.assemble_stabilization_stokes — Method

assemble_stabilization_stokes(
    mesh::MinFEM.Mesh
) -> SparseArrays.SparseMatrixCSC{Float64, Int64}

Returns stabilization matrix for saddle point system in the Stokes equation.

MinFEMFlow.output_end — Method

output_end(verbose::Bool, tsetup::Float64, tsol::Float64)

Writes computation time and summary to console if Stokes solver is verbose.

MinFEMFlow.output_header — Method

output_header(verbose::Bool)

Writes header to console if Stokes solver is verbose.

MinFEMFlow.output_setup — Method

output_setup(verbose::Bool, tsetup::Float64)

Writes setup time to console if Stokes solver is verbose.

MinFEMFlow.print_line — Method

print_line()

Prints a vertical line of width corresponding to the log.

MinFEMFlow.solve_stokesflow! — Method

solve_stokesflow!(fs::FlowSolver)

Returns solution of Stokes flow [u,p] given on given mesh as vector. The inlet velocity function f is applied to the inflow_boundary, which has to be a part of the boundary. On the rest of this boundary no-slip conditions will be applied. Note that the outflow boundary is not given since it is designed to be a free boundary.

Index

MinFEMFlow.FlowSolver

MinFEM.solve!

MinFEM.write_to_vtk

MinFEMFlow.StokesFlowSolver

MinFEMFlow.assemble_incompressibility_stokes

MinFEMFlow.assemble_saddlepointmatrix_stokes

MinFEMFlow.assemble_stabilization_stokes

MinFEMFlow.copy_result!

MinFEMFlow.invalidate!

MinFEMFlow.output_end

MinFEMFlow.output_header

MinFEMFlow.output_setup

MinFEMFlow.print_line

MinFEMFlow.set_boundary_inflow!

MinFEMFlow.set_boundary_noslip!

MinFEMFlow.set_inflow!

MinFEMFlow.set_mesh!

MinFEMFlow.set_result!

MinFEMFlow.set_verbose!

MinFEMFlow.solve_stokesflow!